Modify concern file names
