--- app/controllers/concerns/spree_multi_domain/create_line_item_support.rb
@@ -1,16 +1,0 @@
-module SpreeMultiDomain::CreateLineItemSupport
-  extend ActiveSupport::Concern
-
-  included do
-    prepend(InstanceMethods)
-    rescue_from SpreeMultiDomain::LineItemConcerns::ProductDoesNotBelongToStoreError, with: :product_does_not_belong_to_store
-  end
-
-  module InstanceMethods
-    private
-
-    def product_does_not_belong_to_store
-      render json: { message: I18n.t('spree.errors.products_from_different_stores_may_not_be_added_to_this_order') }, status: 422
-    end
-  end
-end
